reverts usd function and creates a new hook
